fix rbenv install script, core install script

+++ bootstrap.sh
@@ -2,22 +2,40 @@
 
 function install_core {
 	# version control system
-	sudo apt-get install git subversion mercurial
+	sudo apt-get install -qq git subversion mercurial
 
 	# editor
-	sudo apt-get install emacs emacs-goodies-el vim-common
+	sudo apt-get install -qq emacs emacs-goodies-el vim-common
 
 	# python
-	sudo apt-get install python python2.7-dev python-virtualenv python-pip
+	sudo apt-get install -qq python2.7-dev python-virtualenv python-pip
 
 	# node
-	sudo apt-get install nodejs
-	sudo ln -s /usr/bin/nodejs /usr/bin/node
+	sudo apt-get install -qq nodejs
+	if [ ! -f /usr/bin/node ]; then
+		sudo ln -s /usr/bin/nodejs /usr/bin/node
+	fi
+
+	# core util
+	sudo apt-get install -qq tree unzip
+
+	# linm
+	# http://kldp.net/projects/mls/download
+	sudo apt-get install -qq libsmbclient libssl0.9.8
+	local linm_url="http://kldp.net/frs/download.php/4509/linm_0.8.1-1_amd64.deb"
+	local tmp_linm_deb="/tmp/linm.deb"
+	wget -O $tmp_linm_deb $linm_url
+	sudo dpkg -i $tmp_linm_deb
 }
 
 function install_oh_my_zsh {
-	# https://github.com/robbyrussell/oh-my-zsh
-	curl -L http://install.ohmyz.sh | sh
+	if [ ! -d ~/.oh-my-zsh ]; then
+		# https://github.com/robbyrussell/oh-my-zsh
+		curl -L http://install.ohmyz.sh | sh
+	else
+		echo "oh-my-zsh already exist"
+		return
+	fi
 
 	# https://github.com/yonchu/grunt-zsh-completion
 	export OHMYZSH_CUSTOM_PLUGIN_PATH="$HOME/.oh-my-zsh/custom/plugins"
@@ -26,32 +44,49 @@ function install_oh_my_zsh {
 
 function install_rbenv {
 	# https://gorails.com/setup/ubuntu/14.04
-	sudo apt-get install git-core curl zlib1g-dev build-essential libssl-dev libreadline-dev libyaml-dev libsqlite3-dev sqlite3 libxml2-dev libxslt1-dev libcurl4-openssl-dev python-software-properties
+	sudo apt-get install -qq git-core curl zlib1g-dev build-essential libssl-dev libreadline-dev libyaml-dev libsqlite3-dev sqlite3 libxml2-dev libxslt1-dev libcurl4-openssl-dev python-software-properties
 
-	git clone git://github.com/sstephenson/rbenv.git ~/.rbenv/
-	git clone git://github.com/sstephenson/ruby-build.git ~/.rbenv/plugins/ruby-build
+	[ -z "$RBENV_ROOT" ] && export RBENV_ROOT="$HOME/.rbenv"
+
+	if [ ! -d $RBENV_ROOT ]; then
+		git clone git://github.com/sstephenson/rbenv.git $RBENV_ROOT
+	else
+		echo "rbenv already exist"
+	fi
+	
+	local RUBY_BUILD_PATH="$RBENV_ROOT/plugins/ruby-build"
+	if [ ! -d $RUBY_BUILD_PATH ]; then
+		git clone git://github.com/sstephenson/ruby-build.git $RUBY_BUILD_PATH
+	else
+		echo "ruby-build already exist"
+	fi
 
 	# https://github.com/rkh/rbenv-update
-	[ -z "$RBENV_ROOT" ] && export RBENV_ROOT="$HOME/.rbenv"
 	mkdir -p "$RBENV_ROOT/plugins"
-	git clone https://github.com/rkh/rbenv-update.git "$RBENV_ROOT/plugins/rbenv-update"
+	local RBENV_UPDATE_PATH="$RBENV_ROOT/plugins/rbenv-update"
+	if [ ! -d $RUBY_BUILD_PATH ]; then
+		git clone https://github.com/rkh/rbenv-update.git $RBENV_UPDATE_PATH
+	else
+		echo "rbenv-update already exist"
+	fi
 }
 
 case "$1" in
 	"core")
-		echo "Install core"
+		echo "# Install core"
 		install_core
 		;;
 	"oh-my-zsh")
-		echo "Install oh-my-zsh"
+		echo "# Install oh-my-zsh"
 		install_oh_my_zsh
 		;;
 	"rbenv")
-		echo "Install rbenv"
+		echo "# Install rbenv"
 		install_rbenv
 		;;
 	*)
-		echo "unknown"
+		echo "# unknown command"
+		echo "select core, oh-my-zsh, rbenv"
 		;;
 esac
